+++ folio-spring-base/src/main/java/org/folio/spring/utils/FolioExecutionContextUtils.java
@@ -23,7 +23,7 @@ public class FolioExecutionContextUtils {
       var headersCopy = SerializationUtils.clone((HashMap<String, Collection<String>>) context.getAllHeaders());
       headersCopy.entrySet().removeIf(e -> TENANT.equalsIgnoreCase(e.getKey()));
       headersCopy.put(TENANT, List.of(tenantId));
-      log.info("FOLIO context initialized with tenant {}", tenantId);
+      log.debug("FOLIO context initialized with tenant {}", tenantId);
       return new DefaultFolioExecutionContext(folioModuleMetadata, headersCopy);
     }
     throw new IllegalStateException("Okapi headers not provided");