+++ folio-spring-tenant-settings/src/test/java/org/folio/tenant/settings/controller/TenantSettingsControllerTestConfiguration.java
@@ -1,0 +1,12 @@
+package org.folio.tenant.settings.controller;
+
+import org.folio.tenant.settings.config.TenantSettingsProperties;
+import org.springframework.boot.SpringBootConfiguration;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.ComponentScan;
+
+@SpringBootConfiguration
+@ComponentScan(basePackages = "org.folio.tenant.settings")
+@EnableConfigurationProperties(TenantSettingsProperties.class)
+public class TenantSettingsControllerTestConfiguration {
+}